base_script/subtest__net_dropmonitor.sh: new test
This script monitors the system for droped packets and produce a report
of drop locations and counts. The test checks for the expected regexes and
also does a sample count check, which tests whether the number of dropped frames
reported by the script is the same number as the nubmer of dropped frames.