base_script/subtest__futex-contention.sh: new test
The script futex-contention shows a summary of locks, how many times they were contended
and an average time between the beginning and completion of a futex wait. The test checks
whether it matches the expected regexes and also if the number of syscalls reported by the
script puts together the same number as the number of samples.