base_script/subtest__failed-syscalls-by-pid.sh: new test
The script show the summary of syscall errors by pid. Each pid has several syscalls,
which are intented, so we can see under which pid does the syscall belong. This test
checks wheter it matches the expected regexes. In the future we could also add a test
for sample check.